Late night fixes
